requirements.txt
pandas
numpy
pyspark
requests
minio
python-dotenv
